RayCast update

+++ src/logic/PlayerController.h
@@ -41,7 +41,7 @@ private:
 	void updateInteraction();
 public:
 	static glm::vec3 selectedBlockPosition;
-	static glm::vec3 selectedBlockNormal;
+	static glm::ivec3 selectedBlockNormal;
 	static glm::vec3 selectedPointPosition;
 	static int selectedBlockStates;
 	static int selectedBlockId;
